refactor(paths): centralized Windows-to-WSL conversion
Moved drive-path normalization and mount conversion into pi-utils. Both tool path resolution and WSL user-profile discovery now use the same converter, while discovery retains its POSIX fallback.
