+++ packages/utils/src/path.ts
@@ -1,3 +1,17 @@
+import * as path from "node:path";
+
+const WINDOWS_DRIVE_PATH = /^([A-Za-z]):[\\/](.*)$/;
+
+/** Maps an absolute Windows drive path to the drive's default WSL mount. */
+export function windowsPathToWslMount(filePath: string): string | undefined {
+	const normalized = path.win32.normalize(filePath.trim());
+	const match = WINDOWS_DRIVE_PATH.exec(normalized);
+	if (!match) return undefined;
+	const [, drive, rest] = match;
+	const segments = rest.split("\\").filter(Boolean);
+	return path.posix.join("/mnt", drive!.toLowerCase(), ...segments);
+}
+
 const WINDOWS_DRIVE_EXTENDED_PREFIX = /^\\\\[?]\\([A-Za-z]:[\\/].*)$/;
 const WINDOWS_UNC_EXTENDED_PREFIX = /^\\\\[?]\\UNC[\\/]([^\\/]+)[\\/](.+)$/i;
 const WINDOWS_DRIVE_EXTENDED_FORWARD_PREFIX = /^\/\/[?]\/([A-Za-z]:\/.*)$/;
