+++ packages/utils/test/path.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "bun:test";
-import { stripWindowsExtendedLengthPathPrefix } from "../src/path";
+import { stripWindowsExtendedLengthPathPrefix, windowsPathToWslMount } from "../src/path";
 
 describe("stripWindowsExtendedLengthPathPrefix", () => {
 	it("removes drive and UNC extended-length prefixes on Windows", () => {
@@ -14,5 +14,15 @@ describe("stripWindowsExtendedLengthPathPrefix", () => {
 	it("leaves non-Windows paths unchanged", () => {
 		const path = "\\\\?\\C:\\Users\\Shi Xin\\omp.exe";
 		expect(stripWindowsExtendedLengthPathPrefix(path, "linux")).toBe(path);
+	});
+});
+
+describe("windowsPathToWslMount", () => {
+	it("clamps parent traversal at the Windows drive root", () => {
+		expect(windowsPathToWslMount("C:\\..\\Windows\\x")).toBe("/mnt/c/Windows/x");
+	});
+
+	it("rejects paths without an absolute Windows drive", () => {
+		expect(windowsPathToWslMount("/home/me/file.txt")).toBeUndefined();
 	});
 });
